finish rates

+++ migrations/sqls/20221011141809-create-movies-table-up.sql
@@ -3,7 +3,7 @@ CREATE TABLE movies (
     id SERIAL PRIMARY KEY,
     title VARCHAR(255) UNIQUE NOT NULL,
     description text,
-    rate integer NULL,
+    rate float NULL,
     image text,
     user_id bigint REFERENCES users(id) NOT NULL,
     category_id bigint REFERENCES categories(id) NOT NULL
