+++ migrations/sqls/20221011205408-create-rates-table-up.sql
@@ -1,7 +1,7 @@
 /* Replace with your SQL commands */
 CREATE TABLE rates (
     id SERIAL PRIMARY KEY,
-    rate integer NOT NULL,
+    rate float NOT NULL,
     user_id bigint REFERENCES users(id) NOT NULL,
     movie_id bigint REFERENCES movies(id) NOT NULL
 );
